Direct logged in user to correct page if they click login
Modified 'checkPermissions' to 'checkPermissionsSession' which uses the session data rather than the returned user object.
This means we can check permissions outside of the login funciton.
Call the User::checkPermissionsSession function from 'login' in the users_controller to tell where the user should be sent if they are already logged in.